pyproject.toml
[project]
name = "desafio-03-dpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.5.2",
    "jupyter>=1.1.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.49",
]
